ml/pyproject.toml
[project]
name = "sentiment-ml"
version = "0.1.0"
description = "Sentiment classification model for reputation monitoring"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "torch>=2.1.0",
    "transformers>=4.40.0",
    "datasets>=2.18.0",
    "accelerate>=0.27.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "pyyaml>=6.0",
    "tqdm>=4.66.0",
    "evaluate>=0.4.0",
    "ipykernel>=7.2.0",
    "pandas-stubs>=2.3.3.260113",
    "openpyxl>=3.1.5",
    "ipykernel",
    "black",
    "ruff",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "boto3>=1.42.68",
    "sentence-transformers>=5.3.0",
    "hdbscan>=0.8.41",
    "umap-learn>=0.5.11",
    "python-dotenv>=1.2.2",
    "boto3-stubs>=1.42.68",
    "seaborn>=0.13.2",
]


[tool.uv]
dev-dependencies = [
    "ipykernel",
]
